Refactor i3 install script to be re-runnable
Also fix i3lock color setup

+++ install_i3.sh
@@ -9,21 +9,25 @@ echo "This script installs i3, assuming you are running Ubuntu or Manjaro!"
 echo ""
 read -rp "Press enter to start..."
 
-if ! which i3 >&/dev/null; then
-    echo "Installing packages..."
-    if is_ubuntu; then
-        sudo apt-get install -y --no-install-recommends i3 feh xss-lock wmctrl scrot picom dunst rofi pulseaudio-utils playerctl brightnessctl polybar flameshot imagemagick yad libsass1 python3-virtualenv
+echo "Installing packages..."
+if is_ubuntu; then
+    sudo apt-get install -y --no-install-recommends i3 feh xss-lock wmctrl scrot picom dunst rofi pulseaudio-utils playerctl brightnessctl polybar flameshot imagemagick yad libsass1 python3-virtualenv
 
-        # Build i3lock-color
-        sudo apt-get install -y --no-install-recommends autoconf gcc make pkg-config libpam0g-dev libcairo2-dev libfontconfig1-dev libxcb-composite0-dev libev-dev libx11-xcb-dev libxcb-xkb-dev libxcb-xinerama0-dev libxcb-randr0-dev libxcb-image0-dev libxcb-util-dev libxcb-xrm-dev libxkbcommon-dev libxkbcommon-x11-dev libjpeg-dev
+    echo "Building i3lock color..."
+    if ! which i3lock >&/dev/null; then
+        sudo apt-get install -y --no-install-recommends autoconf gcc make pkg-config libpam0g-dev libcairo2-dev libfontconfig1-dev libxcb-composite0-dev libev-dev libx11-xcb-dev libxcb-xkb-dev libxcb-xinerama0-dev libxcb-randr0-dev libxcb-image0-dev libxcb-util-dev libxcb-xrm-dev libxkbcommon-dev libxkbcommon-x11-dev libjpeg-dev libgif-dev
 
         git clone https://github.com/Raymo111/i3lock-color.git
         cd i3lock-color
         ./install-i3lock-color.sh
         cd ..
         rm -rf i3lock-color
+    else
+        echo "... skipping!"
+    fi
 
-        # Clipmenu
+    echo "Building clipmenu..."
+    if ! which clipmenu >&/dev/null; then
         sudo apt-get install -y --no-install-recommends xsel libxcomposite-dev
 
         git clone https://github.com/cdown/clipnotify.git
@@ -37,39 +41,33 @@ if ! which i3 >&/dev/null; then
         sudo make install
         cd ..
         sudo rm -rf clipmenu
-
-        # Older Ubuntu
-        if grep -q "22.04" /etc/os-release; then
-            echo "Manually build i3-gaps..."
-            sudo apt-get install -y --no-install-recommends libcairo2-dev libpango1.0-dev libyajl-dev libxcb-icccm4-dev libxcb-keysyms1-dev libxcb-cursor-dev libstartup-notification0-dev xmlto meson asciidoc
-            git clone https://github.com/Airblader/i3 i3-gaps
-            cd i3-gaps
-            git checkout gaps
-            meson -Ddocs=true -Dmans=true ../build
-            meson compile -C ../build
-            sudo meson install -C ../build
-            cd ..
-            rm -rf build i3-gaps
-        fi
-    elif is_arch; then
-        # Manjaro (libsass is needed by GTK theming tool)
-        sudo pacman -S --needed --noconfirm i3-wm feh wmctrl picom yay dunst rofi clipmenu playerctl brightnessctl polybar scrot xss-lock flameshot imagemagick yad libsass python-virtualenv
-        sudo yay -S --noconfirm --ask 4 --useask --answerclean All --answerdiff None i3lock-color
     else
-        echo "Unsupported OS!"
-        exit 1
+        echo "... skipping!"
     fi
+elif is_arch; then
+    # Manjaro (libsass is needed by GTK theming tool)
+    sudo pacman -S --needed --noconfirm i3-wm feh wmctrl picom yay dunst rofi clipmenu playerctl brightnessctl polybar scrot xss-lock flameshot imagemagick yad libsass python-virtualenv
+    sudo yay -S --noconfirm --ask 4 --useask --answerclean All --answerdiff None i3lock-color
+else
+    echo "Unsupported OS!"
+    exit 1
+fi
 
-    echo "Add sudo privileges to brightnessctl..."
-    sudo chmod +s "$(which brightnessctl)"
+echo "Add sudo privileges to brightnessctl..."
+sudo chmod +s "$(which brightnessctl)"
 
-    echo "Launch clipmenud service..."
+echo "Launch clipmenud service..."
+if [ ! -f ~/.config/environment.d/clipmenud.conf ]; then
     mkdir -p ~/.config/environment.d/
     ln -s "$PWD/i3/clipmenu/clipmenud.conf" ~/.config/environment.d/clipmenud.conf
     systemctl enable --user clipmenud.service
     systemctl start --user clipmenud.service
+else
+    echo "... skipping!"
+fi
 
-    echo "Install rofi themes..."
+echo "Install rofi themes..."
+if [ ! -f ~/.config/rofi/colors/catppuccin.rasi ]; then
     git clone --depth=1 https://github.com/adi1090x/rofi.git
     cd rofi
     ./setup.sh
@@ -79,8 +77,12 @@ if ! which i3 >&/dev/null; then
     sed -i 's|@import .*|@import "~/.config/rofi/colors/catppuccin.rasi"|g' ~/.config/rofi/powermenu/type-1/shared/colors.rasi
     sed -i "s/theme='style-1'/theme='style-3'/g" ~/.config/rofi/powermenu/type-1/powermenu.sh
     sed -i "s|\si3lock|$HOME/.config/i3/scripts/lock.sh|g" ~/.config/rofi/powermenu/type-1/powermenu.sh
+else
+    echo "... skipping!"
+fi
 
-    echo "Install GTK theme..."
+echo "Install GTK theme..."
+if ! compgen -G ~/.local/share/themes/catppuccin*/gtk* >&/dev/null; then
     (
         git clone --recurse-submodules https://github.com/catppuccin/gtk
         cd gtk
@@ -91,8 +93,8 @@ if ! which i3 >&/dev/null; then
         cd -
         rm -rf gtk
     )
-
-    echo "Log into an i3 session to access your new desktop!"
+else
+    echo "... skipping!"
 fi
 
 echo "Configure i3..."
